requirements.txt

streamlit
deepface
opencv-python-headless
pillow
